0.1.31 dynamic styles v1

+++ app/models/sibu/dynamic_style.rb
@@ -24,7 +24,8 @@ module Sibu
       rescue Exception => ex
           logger.error(ex)
       ensure
-        File.delete(css_file_path)
+        File.delete(scss_file_path) if File.exist?(scss_file_path)
+        File.delete(css_file_path) if File.exist?(css_file_path)
       end
     end
